+++ projects/js-packages/charts/src/index.ts
@@ -10,16 +10,23 @@ export { ConversionFunnelChart } from './components/conversion-funnel-chart';
 // Chart components
 export { BaseTooltip } from './components/tooltip';
 export { Legend, useChartLegendItems } from './components/legend';
+export type { LegendValueDisplay } from './components/legend';
 
 // Themes
 export { ThemeProvider } from './providers/theme';
 export { defaultTheme, jetpackTheme, wooTheme } from './providers/theme/themes';
 
 // Global context
-export { GlobalChartsProvider } from './providers/chart-context';
+export {
+	GlobalChartsProvider,
+	useGlobalChartsContext,
+	useGlobalChartsTheme,
+	GlobalChartsContext,
+} from './providers/chart-context';
 
 // Types
 export type * from './types';
 export type * from './visx/types';
+export type { PieChartProps } from './components/pie-chart';
 
 export type { LineStyles, GridStyles, EventHandlerParams } from '@visx/xychart';