+++ projects/packages/connection/src/sso/class-user-admin.php
@@ -19,6 +19,10 @@ use WP_Error;
 use WP_User;
 use WP_User_Query;
 
+if ( ! defined( 'ABSPATH' ) ) {
+	exit( 0 );
+}
+
 /**
  * Jetpack sso user admin class.
  *