+++ projects/packages/boost-core/CHANGELOG.md
@@ -5,13 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.13] - 2025-09-08
+### Changed
+- Internal updates.
+
 ## [0.3.12] - 2025-08-04
 ### Added
 - Add helper to clear all custom transients at the same time. [#44549]
 
 ## [0.3.11] - 2025-06-23
 ### Fixed
-- General: ensure the correct home page is checked when sites are using WP in a subfolder. [#44007]
+- General: Ensure the correct home page is checked when sites are in a subfolder. [#44007]
 
 ## [0.3.10] - 2025-04-28
 ### Changed
@@ -137,6 +141,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Introduce new package. [#31163]
 
+[0.3.13]: https://github.com/Automattic/jetpack-boost-core/compare/v0.3.12...v0.3.13
 [0.3.12]: https://github.com/Automattic/jetpack-boost-core/compare/v0.3.11...v0.3.12
 [0.3.11]: https://github.com/Automattic/jetpack-boost-core/compare/v0.3.10...v0.3.11
 [0.3.10]: https://github.com/Automattic/jetpack-boost-core/compare/v0.3.9...v0.3.10